Library (reading WCS): removed warning for unreasonable LONPOLE
In the previous commit, an extra warning was added to report unreasonable
LONPOLE values. But this was wrong: when its not defined it always gets an
absurd value (even when there isn't any PV matrix!), and default value is
written for it automatically.

With this commit, the warning has been removed.