Fits library: can read output when date-time string ends in Z
Until now, when the date-time string ended in a `Z' the Fits program's
`--datetosec' option wouldn't parse it and complain. This is what is
expected according to the FITS standard. However, unfortunately some FITS
writers don't follow the standard fully and implement the extended w3.org
standard (https://www.w3.org/TR/NOTE-datetime).

According to the `w3.org' standard, a `Z' at the end of the string
signifies that its in UTC (which is the default in FITS). With this commit,
we just ignore the ending `Z' and use the same steps as before.

This feature was suggested by Zahra Sharbaf.