Arithmetic: the interpmetric option is not ignored
Until now, in Arithmetic we were blindly ignoring all command-line options
related to tesselation. This included the `interpmetric' option which is
used by Arithmetic's `interpolate-medianngb' operator!

With this commit, to fix this, all the tesselation options are still
ignored, except this option.

This bug was found by Zahra Sharbaf.

This fixes bug #57921.