Text library: no white-space when a single number is printed on stdout
Until now, when a single number is printed on the standard output, the text
library would print it just like it prints many columns in a text file: it
would add extra white space after the column value. This was very annoying
when used in scripts: the white space needed to be trimmed when the value
is printed in a higher-level structure.

With this commit, it now checks if the output is on printed on the standard
output and if it is only a single number. In this scenario, it will just
use the `gal_type_to_string' function to generate the data's string and
prints that string with no extra white space.