Arithmetic library: more complete sanity checks in arithmetic_size
Until now, the `arithmetic_size' function had the following issues:
  1) The sanity checks weren't complete.
  2) the `arg' dataset wouldn't be freed/not-freed properly.
  3) it would report the length along the C dimension, while astronomers
     expect the FITS order (which is similar to Fortran's dimension
     ordering (the slowest dimension is counted as first).

With this commit, some extra sanity checks were added to the
`arithmetic_size' function of `lib/arithmetic.c', and the error messages of
the existing ones were made more clear for the users. Also, the value of
the `arg' dataset is now immediately written into a `size_t' variable, and
the dataset is freed (or not!) at that point. Finally the output value is
the length along the FITS order.

Also in the documentation, the description of the new `size' operator was
moved after all the operators that deal with a dataset's data. This is
because this operator deals with the dataset's metadata. Later, other
operators that deal with metadata should be added under it. We should also
generally start classifying the operators to be more readable/findable.