Warp: correctly accounting for translation in matrix
Until now, when the warping matrix involved translation, it would take any
affect on the pixel grid beyond the sub-pixel level. It would only correct
the WCS. The cause of this problem was that the final image size and its
starting point were defined based on the differences of the image extremes
(which would cancel out the translation).

With this commit, when the third/sixth elements of the matrix are non-zero,
it will change the output size and correct the output's first-pixel value
(`p->outfpixval').

This fixes bug #57989.