Library(polygon): Added new concave sort utility
The new sort utility can sort both convex and non-convex polygon.
This is based on sorting the points w.r.t a diagonal vector and then
sorting according to their x-coordinates. The function
'gal_polygon_sort_vertex' utilises the other funtions in the file
which all do trivial tasks of splitting and testing the points.