Pointer library: memory-mapped files only written in .gnuastro_mmap
Until now, when a file needed to be memory-mapped, Gnuastro's programs
(through the `pointer.h' library) would first try writing the mmap files in
the `.gnuastro' directory. When it failed it would attempt writing in the
`.gnuastro_mmap' directory. However, `.gnuastro' is also used to store
configuration files (which are hand-written and thus valuable). Mixing the
two types of source (configuration files) and automatically generated
(memory-mapped) files is very problematic.

With this commit, it no longer attempts to write memory-mapped files under
`.gnuastro'. It will only attempt to write them under the `.gnuastro_mmap'
directory.