FITS library: timezone and daylight factored out of date_to_seconds
Until now, in the `gal_fits_key_date_to_seconds' function, we would simply
return the value of the C library's `mktime'. However, `mktime' also
accounts for the host's timezone and daylight saving paramters. So in
effect, the result would be different on different systems (based on their
timezone).

With this commit, we subtract the standard POSIX C library global variables
`timezone' and `daylight' from the output of `mktime'. I tested it on my
system by changing my system's timezone and this fixed the problem.

This bug was found by Zahra Sharbaf.

This fixes bug #57995.