Library(polygon): Added new functions inside polygon.c
Two new functions are added in polygon.c. `gal_polygon_isconvex` is used
to check whether the polygon is concave or convex.
`gal_polygon_isinside` is used to check whether a given point
lies inside the polygon or not. The function `gal_polygon_pin` is
changed to `gal_polygon_isinside_convex`.