Table library: Added an option for concatenating columns of two tables.
Until now, There was no option to concatenate columns of two
tables. With this commit you can concatenate columns of two
tables. One table is given as an must input file and another one
is given to "--catcolumn" option. The column of the table given to
"--catcolumn" option are attached at the back of given input table.