Library (polygon): improved documentation, new functions in NEWS
With the new set of polygon functions defined in the previous commit, it
was necessary to improve the relevant part of the documentation so they
blend in better with the others.

For example the discussion on the definitions/differences of a concave and
convex polygon is now brough to the start of the `polygon.h' subsection and
the new functions were placed in a better logical order
(`gal_polygon_isinside' is now placed before
`gal_polygon_isinside_convex'). This re-ordering was also reflected in
`polygon.c' and `polygon.h'.

Also, Sachin's name was added as a contributing author in the two polygon
library files.