Crop: polygon vertices can now be concave (inner angles >180 deg)
Until now, the `--polygon' option of the Crop program could only work on
convex polygons (where all inner points have an angle less than
180). However, the deep fields of many astronomical surveys
(stacked/coadded images in general), are a concave polygon and it was not
possible to extract them accurately.

With this commit, thanks to the two new functions that Sachin added a few
commits earlier in the polygon library, it is now possible to also extract
convex polygons from an image also. But Concave polygons are not unique, so
it was also necessary to add the new `--polygonnosort' option to crop. With
this option, the polygon vertices are no longer sorted, allowing users to
extract the exact concave polygon that they want.