Table: enable multiple calls to --catcolumn
Until now, only a single file could be column-concatenated with the
`--catcolumn' option. With this commit, this option has been defined as a
list, enabling multiple files to be concatenated with one command.

Also, the `catcolumn' and `catcolhdu' option definitions in `args.h' are
now placed after the previous input option defintions (when the order
doesn't matter, its better to keep the older ones first, for the users its
irrelevant).