Crop: input polygon not sorted by default, must use --polygonsort
Until now, polygon sorting was done by default and the users had to use the
`--polygonnosort' option to disable it. But this is very un-intuitive,
especially when the input is a concave polygon (where the sorted result may
be different). It is just necessary to make sure the polygon's orientation
is counterclockwise.

With this commit, the added functions in the previous commit are used to
reveerse the input polygon vertices into a counter-clockwise orientation by
default (if necessary!), but it will not re-arrange the polygon vertices
beyond that. If the user knows that their points may not be sorted properly
and would like to have the old behavior of sorting the vertices, they can
use the new `--polygonsort' option.

While doing this, I also cleaned up the two newly added functions
(`gal_polygon_is_counterclockwise' and `gal_polygon_to_counterclockwise')
of the previous commit. In particular, since the output of
`gal_polygon_is_counterclockwise' is a binary, the `switch' statement was
replaced with a simple `if' statement.

This completes task #15578.