Table: four new column arithmetic operators to convert RA and Dec
In the previous commit, Kartik introduced two the four new operators and
the necessary library functions. However, the names weren't too descriptive
and have now been changed to `ra-to-degree', `dec-to-degree',
`degree-to-ra' and `degree-to-dec'. Many stylistic corrections were also
made to the original code to make it fit in better with the whole of
Gnuastro and the necessary documentation was added.

In the process, I also noticed that it is much better to specify column in
column arithmetic with a `$' sign (like AWK), not the old `c' character.