Crop: Minor corrections in concave-sort and its usage in Crop
In the previous commit, Sachin implemented the concave sort function as a
library function in Gnuastro and also called it within Crop. Only one issue
was found regarding the check for the warning of sorted concave
polygon. The check had to be done after we sort the polygon vertices in a
final array.

With this commit, that was fixed and some stylistic issues were also
implemented to make the code blend in better with the rest of Gnuastro's
source. The documentation of `gal_polygon_vertices_sort' was also improved
in the book and an example was added.

This completes task #13658.