Book: better explanation of surface brightness limit with an example
Until now, in the discussion of the surface brightness limit, I hadn't
actually explained where the term `sqrt(A/p)' comes from! With this commit,
an explanation is added and even an example set of codes is put so readers
can get a better feeling for this.

This was suggested by Raúl Infante-Sainz.