MakeCatalog: new --forcereadstd for read STD image even if not needed
Until now, MakeCatalog would only read the standard deviation (STD) image
when one of the input columns needed it. However, one of MakeCatalog's
output metadata (the surface brightenss limit) also needs the STD image and
as a result nothing was printed. This could cause confusions.

With this commit, there is a new `--forcereadstd' option that will ensure
that the STD image is read even if no column needs it. Also, when its not
present, instead of not printing anything at all, it now fills the same two
lines with a notice of why there is no surface brightness limit.

Also, while testing on the sample SDSS images, I noticed that the 10^4
scale that we have set for a warning on WCS is a little too strict, so I
loosened it to 10^5.

This was added after a fruitful discussion with Raúl Infante-Sainz.