Library (statistics): STD and sigclip functions fixed for one input
Until now, when the input to the standard deviation and sigma-clipping
operators only had one element, it would sometimes return NaN results
(which is wrong, the standard deviation of a single-element dataset is
0). The reason was that we were simply using the standard equation for the
standard deviation which can be non-zero due to floating point errors.

With this commit, this has been fixed by not using the equation at all when
there is only element. In such cases the standard deviation functions will
just return 0 and the sigma-clipping will just return the four standard
values (1 for number, value for median and mean and 0 for standard
deviation).

This bug was found while working on a project with Zahra Sharbaf.

This fixes bug #58315.