Table: New --inpolygon, --outpolygon and --polygon options added
With the new `--inpolygon' and `--outpolygon' options, it is now possible
to select rows based on their position relative to a polygon (either inside
of it or outside). The `--polygon' option is shared between the two and is
used to specify the actual polygon.

Since we wanted Table's `--polygon' option to be parsed identially to
Crop's, the polygon parsing function of Crop was moved into the options.h
library as `gal_options_parse_colon_sep_csv' which is now used in both Crop
and Table.