All quoted strings now use apostrophe instead of grave accent
Until now, Gnuastro's quoted strings used a grave accent (`) for opening a
quote and an apostrophe for closing it, something like `this'. But as noted
by Rafael Morales, this is not very clear to the eye.

So I had a look at the GNU Coding Standards, and noticed that under the
section "Quote Characters", it also recommends this style: "Although GNU
programs traditionally used 0x60 (`) for opening and 0x27 (') for closing
quotes, nowadays quotes `like this' are typically rendered asymmetrically,
so quoting "like this" or 'like this' typically looks better."

With this commit, this has been corrected in all of Gnuastro, both in the
printed messages and also the comments in the code.

Also, again as suggested by Rafael, to give a unified interface for quoting
strings, all the double quotations have been removed in the printed
messages (in the C source they looked like '\"'). They are now also shown
with an apostrophe.