Fits program and library: better explanation on HDU contents
Until now, when a HDU didn't have a name, we would just print 'n/a'
(not-applicable) for the name column when the Fits program was run on the
file. But this could be confusing for some, leading them to think that
'n/a' is the actual HDU name.

With this commit, when a HDU doesn't have a name, after the column
description, another line is printed, explicity mentioning that 'n/a' means
that there is no name. This doesn't happen for FITS files where all
extensions have a name.

Also, the error message that came when the data is the pimary HDU has been
modified to be more clear about the problem. We recently found out that it
could be interpretted as "the FITS file has problems", while that was not
the point.