Table: properly reads column arithmetic when newline is commented
When using scripts and longer column arithmetic operations, it will often
happen that the author may want to break the line by commenting the
new-line character and going to the next, something like this:

  asttable table.fits --columns='arith $1 $2 + \
                                       $3 /'

Until now, in such a call, Table would crash, complaining that there is no
column called '\' (thinking that '\' is a column name). With this commit
this bug is fixed by parsing all column values before any other
operation. If a backslash is seen that is followed by a new-line, both
characters are replaced with a single space each (which doesn't affect the
interpretation).

This fixes bug #58371.