Book: fixed incorrect call to column number in table's arithmetic
Recently the format of identifying column numers changed in Table's column
arithmetic (to resemble AWK), while it had been corrected in the book, Raúl
Infante-Sainz informed me that one case in the demonstration commands was
still using a 'c' instead of a '$'. It has been corrected with this commit.