Table: new operator to calculate distance on a flat surface
Until now we had a column arithmetic operator to find the distance between
two points on a sphere ('angular-distance'), but it is often necessary to
find the distance on a flat surface also.

To avoid having to write complex code in such cases, with this commit,
Table now has the 'distance-flat' operator. In order to have a unified
naming convention, the spherical distance operator's name was also changed
to 'distance-on-sphere' (as mentioned above, originally it was
'angular-distance').