Library (polygon.h): minor corrections to avoid compiler warnings
While testing for the 0.12 release, I noticed two compiler warnings for the
'min_index' and 'max_index' variables in 'polygon_leftmost_point' and
'polygon_rightmost_point', so I initialized them (with absurd values to
cause a crash if there is problem).

But then I noticed that it is returning an allocated array inside the
function to the outside, so I also moved the allocation of 'point' to the
caller and passed on its pointer to these functions.