MakeCatalog: correcting accounting for ordered columns with no pixels
Until now, when some of MakeCatalog's input labels didn't cover any usable
pixels (for example on the 'values' dataset, all the pixels under a label
were NaN), and a column was requested that required ordering (for example
'--median'), MakeCatalog will crash with this error message.

 gal_data_initialize: dsize[0]==0. The size of a dimension cannot be zero

This happened because at the start of the function to deal with columns
that need ordering ('parse_order_based'), we would immediately try to
allocate the necessary array to keep the object values (to be
sorted). However, in the situation above, the number of elemnts to allocate
was zero and that would cause this problem.

With this commit, instead of immediately allocating the array, first we
check if there are actually any pixels or not. If there aren't we'll just
write a '0' or NaN in the respective column and leave the function (won't
bother going down to the next steps causing this error).

In the meantime, I also noticed that when writing the sigma-clipping
columns for the clumps, we were mistakenly writing in the wrong array
('pp->ci'), it should have been 'ci'! So this potential/future bug is also
fixed with this commit.

This bug was found while working on a project with Raúl Infante Sainz and
Zahra Sharbaf.

This fixes bug #58434.