Library type.h: integers can be read as floats when ending in . or .0
For numerical analysis, it is common practice to simply put a '.' or '.0'
after them to tell the programs to parse them as floating point (for
example when dividing 2 by 3). However, until now, the
'gal_type_string_to_number' function would parse numbers ending in '.' or
'.0' as integers. The only way to force it to parse them as floating point
was to put a 'f' suffix.

So with this commit, the 'gal_type_string_to_number' function (which is in
charge of parsing strings to numbers) will interpret any number that has a
'.' inside of it as a floating point. Note that it has already been parsed
as a number by the C library, so there is no problem in the '.' having
another meaning.

While correcting the documentation, I noticed that we also force a
floating-point type on the 'pow' (to-power) operator of Arithmetic. But
that is again non-intuitive, so a step was added to internally convert the
values to floating point (when they are integers), and free them before
coming out.

This task was suggested by Carlos Allende Prieto.