Library (fits): Timezone correctly used and not subtracting flag
Until now, the function 'gal_fits_key_date_to_seconds' (which was used by
the Fits program's '--datetosec' option, which is used in
'astscript-sort-by-night') would incorrectly subtract the current system
timezone from the output of 'mktime'. But this doesn't take into effect
that the time in the FITS header may not have the same daylight saving as
the run-time.

With this commit, instead of subtracting the local timezone through the
global variable 'timezone', we are subtracting the timezone at the time of
the input.

Also, we were incorrectly subtracting the global variable 'daylight' from
the resulting time!!! This was wrong: 'daylight' is a flag: only having
values of 0 or 1. sSo all times reported until now, may have been off by
one second (depending on the system's timezone!).

Finally, we weren't actually checking until now if 'mktime' has succeeded
or not! With this commit, if it doesn't (and returns '-1'),
'gal_fits_key_date_to_seconds' will return 'GAL_BLANK_SIZE_T', and the Fits
program (which is currently the only program calling this function) will
print an informative error.

This bug was found by Alexey Dokuchaev.

This fixes bug #58455.