Library (wcs): New function to clean floating point errors in WCS
Until now, extremely small values could be printed in the output WCS
structure (for example PC elements on the scale of 1e-17, or extremely
small differences in CDELT elements).

With this commit, the new 'gal_wcs_clean_errors' function is added. By
default, it is called just before writing the WCS by Gnuastro's FITS
library. It will round such values to be equal in the CDELT and to be 0, 1
or -1 in the PC matrix.