Warp: warning is printed when alignment is not correct
Until now, when the North and East axis are not properly oriented
(left-hand orientation), Warp would simply fail to make the correct
rotation and wouldn't give any warning. It has been recognized as bug

With this commit, it prints a warning when such cases may occur. We still
haven't had time to fix it, but atleast the warning will tell the users how
to fix it.

Also, Warp now prints the name of its output file.