Library (wcs) and Fits program: converting between WCS distortions
The WCS standard has historically had several types of distortions and in
many scenaios it is necessary to convert between them. With this commit, an
internal 'wcsdistortion' library has been added to Gnuastro for the job and
some high-level functions have been added to 'wcs.h' to make it easy to
interact with the low-level features.

To enable easy command-line access to the WCS distortion conversions the
'--wcsdistortion' option has been added to the 'astfits' program. When
given this function, the program will convert the input HDU's WCS
distortion into a the desired type. Currently only the conversion between
TPV to SIP is supported.