Fits: --wcsdistortion with no data HDUs, no crash on old WCSLIB
Until now, when the input FITS header given to the '--wcsdistortion' option
didn't have any data, it would crash, complaining about this! But WCS
distortions are may be stored in FITS HDUs that don't have any data (and
infact this is what happens with astrometry.net for example). Hence it was
necessary to account for this scenario.

To enable this, with this commit, a library function 'gal_wcs_write' was
added to only write the WCS-related keywords in a data-free FITS header and
the necessary checks were placed to avoid reading data when there isn't
any.

In the process (on the same branch that this commit was "rebase"d from),
the following corrections are also made:

 - gcc allows 2D matrix declaration and initialiasation in the form
   `array[i][j]={0}` but clang raises warning for they same. To provide
   uniformity we used nessted loops to initialise these 2D matrices.

 - Until now we were assuming that the user's 'wcsprm' structure has an
   'mjdref' element, but after testing on a system with WCSLIB 5.17, we
   noticed that this element is a relatively recent addition to the
   'wcsprm' structure (it wasn't present in that version). So with this
   commit, a configure-time check has been added so if the host's WCSSLIB
   doesn't have this structure, it isn't used.

 - On WCSLIB 5.17, we also noticed that 'dpkeyd' also doesn't exist, thus
   causing an error during compilation. We thus don't use it any more. It
   was mainly used for extra error check and after a discussion was
   scraped.

 - Until now, we always assumed that WCSLIB has the 'dis.h' header for
   distortion-related operations. But this is not always the case and older
   versions of WCSLIB don't have it. This could cause problems in the
   build. With this commit, configure-time checks have been added to see if
   'dis.h' is present on the host or not. If it isn't the respective
   functions will print an informative error and abort. Also a warning will
   be printed at the end of the configuration step.