Fits: default size with --wcsdistortion=SIP and no input data
Until now, when the Fits program was called with '--wcsdistortion=SIP'
(from a TPV-based WCS) and the input HDU didn't have any data, it would
crash, complaining that a default size is necessary.

With this commit, a set of steps have been added in the function called by
'--wcsdistortion' to allow a default size when the input HDU has no
size.

In the process, a bug was fixed in reading the keywords by
`wcsdistortion_get_tpvparams`. The problem was fixed by reading directly
from 'disseq' parameter in wcs, which also made its reading style similar
to that of `wcsdistortion_get_sipparams`.