Library(polygon): Fixed wrong ordinand index array during sorting
During concave sorting of points, the ordinand index in
`gal_polygon_verices_sort` was a bit wrong due to points being used from
the partially sorted arrays A and B, which ideally should have been the
user given `vertices` array.

It fixed now and no changes have to be made on the function prototypes
and hence no major changes have to be done for the options that uses it.