Warp: not allowing --translate and --centeroncorner together
Until now, just before finding the size of the final warped image, we would
see if there is a translation element in the matrix. If there was a
translation, then the size of the output would be adjusted to account for
the translation. This is required for normal "translations" (so the image
"moves" compared to the first pixel). However, the '--centeroncorner'
option is also a translation, but this adjustment is not desired there and
will result in incorrect output sizes.

With this commit, when the '--centeroncorner' option is called, the
adjustment is not made, and to avoid unexpected problems, it won't allow
calling the '--translate' and '--centeroncorner' options together. Since
the centering is inherently a translation, a solution is suggested in the
error message.

This bug was found by Marjan Akbari.

This fixes bug #58696.