CosmicCalculator: new --listlinesatz option for wavelengths at z
Until now, CosmicCalculator would just print the observed wavelength of one
line at the given redshift with '--lineatz'. But sometimes a more general
view of the shifted spectra can be useful. However, the '--listlines'
option only printed the lines at rest frame.

With this commit, the newly added '--listlinesatz' option does the same
thing as '--listlines' (infact is uses the same function too!), but instead
of printing the restframe wavelength, it prints the observed wavelength. It
very convenient to match/demonstrate spectra for example, when one line is
known/recognized.

Also, with the number of spectral-line features in CosmicCalculator
increasing, I moved their documentation in the book to a new and dedicated
sub-section.