Library (speclines.h): Added SiIII, OIII, CIV and NV and Lyman-series
While reading Feltre et al. 2020 (arXiv:2007.01878), I noticed that we are
missing the first set of lines in the list of Gnuastro's pre-defined
spectral lines. While adding them, I thought it will be good for
completeness if the rest of the Lyman series is also added.

Following the new convention of calling OIII's two doublet's as '_VIS' (for
the visible one) and '_UV' (for the ultra-violet one), the same convention
was applied for HeII.