Table: informative error message for wcstoimg and imgtowcs
These two column arithmetic operators need the same number of operands as
the dimensions in the input WCS structure. However, until now, when the
proper number of columns wasn't given to them, they would just complain
with "not enough operands" which can be really confusing in practice (it
happened for myself! I was using these operators in an IFU cube, but since
I was so focused on the RA and Dec, I forgot about the spectral dimension
and was so confused by this error message that I had to come to the code to
find out the real cause!!!).

With this commit, a new optional argument has been given to the
'arithmetic_stack_pop' function (which is in charge of popping operand) for
this job! When it isn't NULL, the given string will be printed after the
main error message. Currently its only active for these two operators (we
can easily add one for the rest if they need it).