Warp: informative error message when input is a 3D cube
Until now, when the input to Warp was a 3D cube, it would produce an
unreasonable output (because it is only designed for 2D images). Warp
currently only complains about 3D images with the '--align' opertion.

The new task #15729 has been defined for this job, but until it is done, we
need to inform the users that Warp is not yet designed for 3D
operations. So with this commit, an informative warning message is now
printed and Warp is aborted.

This bug was found by Leindert Boogaard.

This fixes bug #58774.