Table: column concatenation with only certain columns
Until now, when doing column concatenation (appending), all the columns in
the specified file would be appended to the output table. This was very
annoying when only one column is needed for example. Also the names of the
column were unchanged (causing many situations where the same name would be
used for more than one column).

With this commit we now have the '--catcolumns' option which can be used to
limit the columns that will be appended. Also, by default the names of all
appended columns will be appended with a '-N' (where 'N' is the file
counter hosting the column). Also the old '--catcolumn' option (which was a
file name) has been renamed to '--catcolumnfile' to avoid confusion.

With these features, it is now very easy to add multiple measurements (for
examples catalogs generated by MakeCatalog but from different filters, that
all have the same column names and some redundant columns) into one.