Table: --metaupdate, new option to update column meta data
Until now, after column arithmetic or column concatenation, the new columns
in the output table didn't have good metadata and the users were forced to
use 'astfits' to update the column names, and units and comments. This was
very inconvenient and was thus easily forgotten or ignored. But metadata
are very important and without them the usability of the table will be very
limited (even the author will forget!).

With this commit, a new '--metaupdate' option has been added that will
allow updating column metadata in the same command that generates the
table.