Book: better explanation of pre-convolved stars in first tutorial
In the first tutorial, we build four stars (as a point source) in the model
image. But hadn't really explained the fact that they only cover a single
pixel and this caused confusion for new readers: when you open the image in
DS9 and look at the full image at once, you don't see the four stars.

With this commit, this important issue is clarified a little more to
hopefully avoid this confusion for future readers.

This issue was mentioned by Samane Raji.