developer-build: number of parallel jobs calculated at run time
Until now, the 'developer-build' script used a hard-coded number for the
number of parallel jobs. But this isn't generally good, because on
different systems, we may need a different number of threads.

With this commit, it now reads the number of threads from the system (using
'nproc') when no value has been given as an option.