Segment: accounting for blank pixels when checking detection map
The input detection map in Segment is interpreted as a signed 32-bit
integer type. So when it has blank values, they are stored as the smallest
possible number in this format (a huge negative number). However, the
checks at the start of Segment don't account for this, so when there are
blank elements in the detection map, Segment crashes, by saying that there
are negative pixel values in the detection map (which should be only
positive).

With this commit, if the value is negative, AND it is equal to the standard
blank value for this type, then it doesn't crash any more. fortunately all
later parts of Segment can account for blank pixels, the problem was only
in the check at the start.

This bug was reported by Zahra Sharbaf.

This fixes bug #58833.