Library (interpolte.h): Min and Max in nearest neighbor
Until now, the only operator for nearest neighbor interpolation was the
median. But the low-level function was easily expandable to any type of
statistic!

So with this commit, we are now adding minimum and maximum also (for
example using maximum of nearest neighbors). This was based on the need to
interpolate between the blank pixels of saturated stars to avoid
segmentation complications while working on a project with Zahra Sharbaf.