Crop: floating point errors accounted, when checking pixel scale
In WCS mode, Crop can take multiple images and stitch them together. To do
this, it needs to make sure that all images have the same pixel scale. But
until now, it used a simple equality check to see if the pixel scales are
the equal. However, floating point errors can create problems in this
check, causing a crash in Crop (which will complain about the images not
having the same pixel scale).

To fix this, instead of using a simple equal, I changed the check to see if
the difference between the two pixel scales is larger than floating point
errors (roughly 1x10^{-10}).

This bug was reported by Joanna Sakowska.

This fixes bug #58835.