Book: fixed typo in dilate operator of Arithmetic
Prior to this commit, the explanation of the dilate operator said
"Erosion is simply a flipping of all background pixels..."

"Erosion" should be "Dilation"

With this commit, I have made this change.  I also improved the
consistency in mentioning the background and foreground values
between the erode and dilate operators.