Table: --colmetadata is new name for recently added --metaupdate
We recently added the '--metaupdate' option to add/update column metadata
to the Table program. But it isn't always a case of updating meta data,
infact based on early experience, it will mostly be used for columns that
have no metadata.

With this commit, to avoid confusing new users, the name of the new option
is now called '--colmetadata'.