Book: general tutorial discusses merging three catalogs also
Until now, in the general tutorial we were only showing how to add the
magnitude of another filter into another one. For some users, this had
caused the confusion that the column concatenation feature is only for two
columns, and they were making many tables from two catalogs instead of a
single table that merges all catalogs.

So with this commit, in the general tutorial, we also process the F125W
image (Crop+NoiseChisel+Segment), and after talking about merging of two
catalogs, we discuss three catalogs (helping the reader into the logical
conclusion of extending to many).