Library (txt.h & fits.h): Properly accounting for blank strings
Until now, the blank value specified for plain-text numeric columns had to
be a number in the same type. This was inconvenient because we would have
to specify unique numbers based on the column type.

With this commit, its now possible to specify strings as blank values of
numeric columns. With this feature, we can now specify the same string for
all numeric column types and just give them in the column metadata.

While doing this, I also found a bug in writing FITS tables: until now for
non-standard FITS types (int8_t, uint16_t, uint32_t and uint64_t), we would
simply write the blank value in the TNULL keyword. However, according to
the FITS standard, TNULL is checked before TZERO or TSCAL are applied to
these non-standard data types. As a result, when reading from that FITS
table, the blank elements aren't specified as blank any more! So this has
also been corrected.

This bug was reported by Samane Raji.

This fixes bug #58901.