Library (txt.h): string columns won't force extra \0 in line
Until now, when reading plain-text tables, we would generally read tokens
in the same allocated line that we had read from the file. So for strings,
we had to put a '\0' after the specified width. This would cause problems
in a table like below, where the first character of the third column ('3')
was being set to '\0' and thus lost. Note that I am using '%' instead of a
hash to keep the comment lines in the commit message.

% Column 1: ID [counter, u8] Counter
% Column 2: NAME [name, str4] a name
% Column 3: Value [km, f32] A value
1 AAAA3.14
2 BBBB3.15

To fix this problem, instead of reading the tokens from the line, we
allocate space for each token, and copy the token's string there. This
fixes the problem mentioned above, because we now allocate separate space
and have a byte to put the '\0'.

This bug was found by Samane Raji.

This fixes bug #58898.