Library (fits.h): corrected TNULL for non-standard types in tables
We recently corrected the writing of TNULL while reading a FITS table for
non-standard integer types, but a similar correction was necessary when
writing a FITS table for these types.

With this commit, a new function has been defined to easily allocate the
necessary value (and correct for the non-standard types).