Library (fits.h): no blanks in writing standard FITS numeric tables
Until now, when writing FITS tables we would simply write Gnuastro's blank
value into the 'blank' pointer that is passed to CFITSIO. But this would
cause an error (which is not expected as far as I understand from the
definition of TNULL in the FITS standard).

With this commit, for these types, we now remove the blank value (and its
pointer) because they are ultimately redundant for these types. This fixed
the problem in cases that we tested.