Statistics program and library: new 2D histogram feature added
In many contexts, a 2D histogram is necessary to study the behavoir of two
features in a large distribution and a scatter plot is usually not enough.

With this commit, the 'statistics.h' library now has the new
'gal_statistics_histogram2d' function and it is used in the Statistics
program to get the necessary columns to draw a 2D histogram. A new section
has also been added in the book, describing 2D histograms that also
contains a fully working minimal LaTeX source to draw the result.