Library (wcs.h): writing CD matrix when TPV distortion present
Until now, Gnuastro would simply follow WCSLIB's default output of writing
the PC+CDELT matrices into the header keywords. However, we recently found
out that the PC+CDELT matrix (when CDELT is not equal to 1) will cause
problems in WCSLIB's TPV conversion. After contacting Mark Calabretta, we
found out that this is a missing issue in WCSLIB and that it will be fixed
in the next version of WCSLIB (7.4, to be released).

This is the full reply of Mark: "the problem was that TPV assumes that
CDi_ja is used to define the linear transformation. (Consequently, the
independent variables of the TPV polynomial are intermediate world
coordinates rather than intermediate pixel coordinates. That's the wrong
way to do it, but can be excused as an historical anomaly.) Thus, unlike
other sequent distortions, if PCi_ja is used instead of CDi_ja, then
CDELTja must be applied before the TPV polynomial. WCSLIB wasn't doing
that".

Even though it will hopefully be fixed in WCSLIB 7.4, it will take a long
time for the community to transition to it (maybe +3 years or so!). So it
is safer if we also implement a correction in Gnuastro and avoid problems
for users who have older WCSLIB versions.

With this commit, if a TPV distortion is present, we'll multiply the CDELT
matrix with the PC matrix (so effectively CDELT=1), then we'll rename the
PC matrix keywords to CD. Because using the CD matrix is non-standard for
Gnuastro, an explanation is added just after the WCS keywords, thoroughly
explaining why we were forced to use the CD matrix.

This fixes bug #58974.